added time stamp w/ date
added a timeSent value to all messages sent by a user containing the date and time (formatted to English standard - dd/mm/yyyy) the message was sent. this is quite over the top in terms of how many lines it took but I couldn't see a better (at least better + readable) way of doing it.